수정 FourOperation -> Operator
